spark/requirements.txt
pyspark
psycopg2-binary
kafka-python
python-dotenv